requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
revChatGPT==4.0.3
openai==0.27.2
asgiref==3.6.0
pymongo==3.12.1